SI-4694-Extend-VMEM-areas-beyond-4-GB (#20)
* SI-4694 Adapt function interfaces to 64 bit addressing

* SI-4694 Adaptions to 64 bit VMEM for 32 bit CPUs

* SI-6494 Use fallback memcpy for vmem_ram areas
Originally because C does not allow initialization of a 64 bit variable from a 32 bit pointer

* SI-4694 Updated the API to use uint64_t types instead of intptr_t and uint64_t

* SI-4694 Extended the param_t structure with a 64-bit vaddr field to be used in case of mapping the parameter to VMEM storage. Changed the len/size from uint64_t to uint32_t since all memory copy operations in the VMEM domain can only be uint32_t since they involve a physical address.

* SI-4694 Removed extra vmem = NULL from param.h define. Changed the alignment of the param_t types to be on a 8-byte boundary since the new vaddr field is 64-bits. Changed the vmem_t to be 8-byte aligned because the vaddr is a 64-bit usigned integer. Extended the vmem_calc_crc32() be be able to calculate CRC32 on a length greater than 32-bits, now 64-bits. Changed the VMEM server version to be 3, with 64-bit support for the size field. The VMEM server now handles 64-bit lengths in both upload, download and crc32 operations.

* Corrected a typecast of the addr element of param_t

* SI-4694 64-bit support for the VMEM clients (upload, download, crc32 and vmem). Small modifications to vmem_cpy() to fit the wpeek and wpoke commands. Be able to support 4GB of file upload in VMEM.

* SI-4694 Corrected some printf() format specifiers to fit

* SI-4694 Updated the vmem file and mmap drivers to support new 64-bit API.

* SI-4694 Fixed portability issue with format specifiers in printf() in vmem_client.

* SI-4694 Fixed issue with vmem_block not handling the virtual address correctly.

* SI-4694 Fixed compiler warning of uninitialized variable in vmem_block.c

* SI-4694 Implemented a new version of VMEM list transfer for version 3, which collects the vmem response from packets of size VMEM_SERVER_MTU's. Now the VMEM name can be up to 16 characters

* SI-4694 VMEM list server now splits the response up into seperate MTU's for version 3.

* SI-4694 Fixed a potential error in the VMEM server list response.

* SI-4694 Modified the vmem_read() and vmem_write() to specifically use the read amd write methods of the vmem structure found. This is because the to and from addresses are being truncated to 32-bits thru the normal vmem_cpy() method. And this may result in collisions between physical and virtual addresses when searching for a vmem match.

* Reduce nesting in VMEM_SERVER_LIST v3

* SI-4694 Updated code after review.

* SI-4694 Corrected code after review

---------

+++ include/vmem/vmem_mmap.h
@@ -11,8 +11,8 @@ typedef struct {
 	char * filename;
 } vmem_mmap_driver_t;
 
-void vmem_mmap_read(vmem_t * vmem, uint32_t addr, void * dataout, uint32_t len);
-void vmem_mmap_write(vmem_t * vmem, uint32_t addr, const void * datain, uint32_t len);
+void vmem_mmap_read(vmem_t * vmem, uint64_t addr, void * dataout, uint32_t len);
+void vmem_mmap_write(vmem_t * vmem, uint64_t addr, const void * datain, uint32_t len);
 
 #define VMEM_DEFINE_MMAP(name_in, strname, filename_in, size_in) \
 	static vmem_mmap_driver_t vmem_mmap_##name_in##_driver = { \
